Move number parsing tool to utility module
